refactory v1 - Test Drive, Users, Auth, Car

+++ src/main/java/com/evdealer/ev_dealer_management/car/repository/DimensionRepository.java
@@ -7,5 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface DimensionRepository extends JpaRepository<Dimension,Integer> {
+public interface DimensionRepository
+        extends JpaRepository<Dimension, Long> {
 }
